+++ .env
@@ -1,0 +1,4 @@
+name = "v i s i o n"
+api_id = 12948706
+api_hash = "072bc1f28487b01e6e583d6c903f318b"
+bot_token = "7758312468:AAGTE060OxWYpf2pK8TRkGeRSRgkKKQgBfk"
