--- bot/__init__.py
@@ -1,1 +1,0 @@
-from .bot_instance import client
