--- client_handlers/__init__.py
@@ -1,3 +1,0 @@
-active_handlers = [
-
-]
