--- client_handlers/base/__init__.py
@@ -1,10 +1,0 @@
-from .handlers import BaseHandler, request_type, Client
-from pyrogram.filters import command, all_filter, regex, create as create_filter
-from pyrogram.handlers import MessageHandler, CallbackQueryHandler
-
-
-__all__ = [
-    "BaseHandler", "request_type", "Client",
-    "command", "all_filter", "regex", "create_filter",
-    "MessageHandler", "CallbackQueryHandler",
-]
