--- env_example
@@ -1,4 +1,0 @@
-name = ""
-api_id = -1
-api_hash = ""
-bot_token = ""
